refactor(*) integrate with `aws-lambda` (#6)
This commit integrates the `liamp` plugin with the `aws-lambda` plugin.

The `liamp` plugin will be renamed back to `aws-lambda`, stay in its
own external repo, and then the upstream changes would be to remove
Lambda there, and update the rockspec to include this external one.

We're bumping from 0.2.0 to 3.0.0 to supersede the `aws-lambda` plugin
in the Kong/kong open source repository (which was already at 2.0.0).