feat(*) configurable body base64 encoding
As not every lambda function uses base64 encoded bodies, now it's
possible to completely disable this body encoding. This makes it easier
for users who are migrating from an environment where lambda functions
are not base64 encoded to Kong.

The default behavior still is encoding the requests and the responses.